+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Multiple Browsers can be created without one affecting the other @raycardillo
+
 ### Added
 
 ### Changed